+++ x/emissions/keeper/msgserver/msg_server_worker_payload.go
@@ -256,6 +256,8 @@ func verifyAndInsertForecastsFromTopForecasters(
 // A tx function that accepts a list of forecasts and possibly returns an error
 // Need to call this once per forecaster per topic inference solicitation round because protobuf does not nested repeated fields
 func (ms msgServer) InsertWorkerPayload(ctx context.Context, msg *types.MsgInsertWorkerPayload) (*types.MsgInsertWorkerPayloadResponse, error) {
+	sdkCtx := sdk.UnwrapSDKContext(ctx)
+	blockHeight := sdkCtx.BlockHeight()
 	_, err := sdk.AccAddressFromBech32(msg.Sender)
 	if err != nil {
 		return nil, err
@@ -264,8 +266,6 @@ func (ms msgServer) InsertWorkerPayload(ctx context.Context, msg *types.MsgInser
 	if err != nil {
 		return nil, err
 	}
-
-	blockHeight := sdk.UnwrapSDKContext(ctx).BlockHeight()
 
 	if err := msg.WorkerDataBundle.Validate(); err != nil {
 		return nil, errorsmod.Wrapf(types.ErrInvalidWorkerData,
@@ -416,7 +416,7 @@ func (ms msgServer) InsertWorkerPayload(ctx context.Context, msg *types.MsgInser
 		if err != nil {
 			return nil, err
 		}
-		_, topNInferer := actorutils.FindTopNByScoreDesc(moduleParams.MaxElementsPerForecast, latestScoresForForecastedInferers, forecast.BlockHeight)
+		_, topNInferer := actorutils.FindTopNByScoreDesc(sdkCtx, moduleParams.MaxElementsPerForecast, latestScoresForForecastedInferers, forecast.BlockHeight)
 
 		for _, el := range forecast.ForecastElements {
 			if !seenInferers[el.Inferer] && topNInferer[el.Inferer] {